api/requirements.txt
tensorflow>=2.16.1
fastapi>=0.135.0
uvicorn[standard]
python-multipart
pillow
numpy>=2.0.0
matplotlib